LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/privacy/scrollbar-thumb-images-ignored.https.sub-expected.txt


FAIL drawElementImage does not draw cross-origin scrollbar thumb images assert_array_equals: Cross origin should not draw expected property 0 to be 0 but got 255 (expected array [0, 0, 255, 255] got object "255,0,0,255")
